+++ vite.config.ts
@@ -1,6 +1,5 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
-import { resolve } from 'path'
 
 const host = process.env.TAURI_DEV_HOST
 
@@ -20,14 +19,6 @@ export default defineConfig({
       : undefined,
     watch: {
       ignored: ['**/src-tauri/**'],
-    },
-  },
-  build: {
-    rollupOptions: {
-      input: {
-        main: resolve(__dirname, 'index.html'),
-        popout: resolve(__dirname, 'popout.html'),
-      },
     },
   },
 })